Use 16-bit color depth. Fix fill color for credits screens.

+++ doom64/cfb.c
@@ -8,7 +8,7 @@
 #include <ultra64.h>
 #include "i_main.h"
 
-u32 cfb[2][SCREEN_WD*SCREEN_HT]; // 8036A000
+u16 cfb[2][SCREEN_WD*SCREEN_HT]; // 8036A000
